refactor(component): move subcomponents to React adapter
Subcomponent realization presumes a coarse-re-render host (bind a render fn
to its owner, re-invoke on change) - a React-family assumption, not part of
the agnostic Component contract. Lift the walk into @expressive/react, where
it calls useHook/watch directly, and drop the subcomponents/Bind exports from
@expressive/component. The core's public surface is now Component + types.

Also install the React internals carve-out (updater/refs/_reactInternals/
_reactInternalInstance) as one-time setter-only accessors on the prototype
instead of a per-construct loop, so instances only materialize a
non-enumerable own property if React actually assigns one.
